Add files via upload
As part of my academic and practical learning in Python programming and computer vision, I undertook a project to develop a vehicle speed detection system. The objective of the project was to implement a real-time application that could accurately estimate the speed of vehicles passing in front of a camera.The Vehicle Speed Detection project provided hands-on experience in computer vision, video processing, and data analysis using Python. By implementing the speed estimation algorithm and handling real-world challenges like camera calibration, the project honed my skills in problem-solving and algorithm design.
The project's successful completion reinforced my passion for computer vision and its applications in real-world scenarios. It also showcased my ability to integrate various Python libraries and technologies to create practical solutions to complex problems.